form modif and clean css
