+++ public/assets/css/form.css
@@ -1,17 +1,7 @@
-section {
+.content-form {
     height: 100vh;
-    background-image: var(--bg-texture);
-    font-family: var(--bs-font-family);
 }
 
-h2 {
-    font-family: var(--bs-font-family);
-}
-
-button:hover{
-    color: var(--bs-success-rgb);
-}
-
-li {
+list-errors {
     list-style: none;
 }
