+++ src/routes.php
@@ -15,5 +15,5 @@ return [
     'items/show' => ['ItemController', 'show', ['id']],
     'items/add' => ['ItemController', 'add',],
     'items/delete' => ['ItemController', 'delete',],
-    'contact' => ['FormController', 'index',['answer']],
+    'contact' => ['ContactController', 'index',['answer']],
 ];
